BUG: User customization of the notification sound needs $TARGET_USER set by messageToRemoteUser (after 76ed818)
messageToRemoteUser → executeOnUserSystemThroughServiceAccount -> sudoWithUnixhome → messageToDesktopUsers
That means that the command actually is invoked by root, so the notification sound file would have to be placed in their home directory; the user's (which is the target of the customization) is ignored.
Fortunately, as root we're able to access any user's home directory, we just need to know who is the target user.
messageToRemoteUser now passed this information along via $TARGET_USER.

(This corrects commit 76ed81842d8b5df3a3bf06e49d9350157c9243a2)