clearNotifications: Robustness: Send a dummy message to obtain the current ID
If we've run out of sync with the persisted last notification ID, we can (again assuming sequential ID numbering) obtain the current ID by sending a dummy message. notify-send doesn't (yet) have a way to obtain the ID, so we need to use the same lower-level gdbus API that we're using to clear, anyway. Source: https://superuser.com/a/1593924/117820